Add slice support for FileRegion.
Motivation:
Let codec such as HTTP/2 can write FileRegion chunk by chunk.

Modifications:
Add a transferIndex which is similar to readerIndex in ByteBuf for FileRegion.
Add two transferBytesTo methods which are similar to readBytes in ByteBuf for FileRegion.
Add slice and transferSlice methods which are similar to slice and readSlice in ByteBuf for FileRegion.
Add a SlicedFileRegion.
Add a channel method for FileRegion which returns the underlying FileChannel.
Make the return value of position to be the offset related to the FileChannel returned by channel method. And for DefaultFileRegion, the return value of the original position method is the offset related to the underlying FileChannel, so this does not break compatibility.
Add SlicedFileRegionTest
The native transport has not been modified yet so it can not accept the SlicedFileRegion right now, will address it in another PR.

Result:
Passed all FileRegion related testsuites.